Move most of the bussines logic to a helper class that will soon be gone when Async will be in place.
